Fix imports + move fixtures
