+++ decompiler/pipeline/controlflowanalysis/__init__.py
@@ -1,5 +1,5 @@
 from .expression_simplification import ExpressionSimplification
 from .instruction_length_handler import InstructionLengthHandler
+from .loop_name_generator import LoopNameGenerator
 from .readability_based_refinement import ReadabilityBasedRefinement
 from .variable_name_generation import VariableNameGeneration
-from .loop_name_generator import LoopNameGenerator
