+++ decompiler/pipeline/controlflowanalysis/variable_name_generation.py
@@ -4,15 +4,7 @@ from enum import Enum
 from typing import List, Optional, Set
 
 from decompiler.pipeline.stage import PipelineStage
-from decompiler.structures.pseudo import (
-    CustomType,
-    Float,
-    GlobalVariable,
-    Integer,
-    Pointer,
-    Type,
-    Variable,
-)
+from decompiler.structures.pseudo import CustomType, Float, GlobalVariable, Integer, Pointer, Type, Variable
 from decompiler.structures.visitors.ast_dataflowobjectvisitor import BaseAstDataflowObjectVisitor
 from decompiler.task import DecompilerTask
